refactor: adapt to the upstream changes
